+++ lib/riaction/iactionable/objects/identifier.rb
@@ -4,6 +4,14 @@ module IActionable
       attr_accessor :id
       attr_accessor :id_hash
       attr_accessor :id_type
+      
+      def to_hash
+        {
+          "ID" => @id,
+          "IDHash" => @id_hash,
+          "IDType" => @id_type
+        }
+      end
     end
   end
 end