+++ lib/riaction/iactionable/objects/level.rb
@@ -10,6 +10,15 @@ module IActionable
         @level_type = IActionable::Objects::LevelType.new(key_values.delete("LevelType"))
         super(key_values)
       end
+      
+      def to_hash
+        {
+          "Name" => @name,
+          "Number" => @number,
+          "RequiredPoints" => @required_points,
+          "LevelType" => @level_type.to_hash
+        }
+      end
     end
   end
 end