+++ lib/riaction/iactionable/objects/level_type.rb
@@ -3,6 +3,13 @@ module IActionable
     class LevelType < IActionableObject
       attr_accessor :key
       attr_accessor :name
+      
+      def to_hash
+        {
+          "Key" => @key,
+          "Name" => @name
+        }
+      end
     end
   end
 end