+++ lib/riaction/iactionable/objects/profile_summary.rb
@@ -8,9 +8,9 @@ module IActionable
     class ProfileSummary < IActionableObject
       attr_accessor :display_name
       attr_accessor :identifiers
-      attr_accessor :points
-      attr_accessor :recent_achievements
-      attr_accessor :rank
+      attr_accessor :points 
+      attr_accessor :recent_achievements # not always present
+      attr_accessor :rank # not always present
       
       def initialize(key_values={})
         @identifiers = extract_many_as(key_values, "Identifiers", IActionable::Objects::Identifier)
@@ -19,6 +19,19 @@ module IActionable
         
         super(key_values)
       end      
+      
+      def to_hash
+        hash = {
+          "DisplayName" => @display_name,
+          "Identifiers" => @identifiers.map{|identifier| identifier.to_hash}
+        }
+        unless @points.nil?
+          hash["Points"] = @points.kind_of?(Array) ? @points.map{|point| point.to_hash} : @points
+        end
+        hash["RecentAchievements"] = @recent_achievements.map{|recent_achievement| recent_achievement.to_hash} unless @recent_achievements.empty?
+        hash["Rank"] = @rank unless @rank.nil?
+        hash
+      end
     end
   end
 end